booking

+++ src/pages/manager/landlord/[id]/booking/index.tsx
@@ -11,6 +11,7 @@ import AddBooking from './addBooking';
 
 type Props = {};
 
+
 const Booking = (props: Props) => {
   const [open, setOpen] = useState(false);
   const onCloseModal = () => setOpen(false);
@@ -20,15 +21,16 @@ const Booking = (props: Props) => {
   const router = useRouter();
   const param = router.query;
   const id = param.id;
+  const [listRooms, setListRooms] = useState<any>();
 
   const {
     register,
     handleSubmit,
     formState: { errors },
   } = useForm<any>();
-  const [listBookings, setListBookings] = useState<any>();
+  const [listBookings, setListBookings] = useState<any>({});
   console.log(listBookings);
-  
+
   useEffect(() => {
     const getListBooking = async () => {
       const { data } = await listBooking(userData, id);
@@ -37,40 +39,69 @@ const Booking = (props: Props) => {
     getListBooking();
   }, []);
 
-  const onCreateRoom = async (data: any) => {
-    const newData = { ...data, userData: userData };
-    await createBookingRoom(newData)
-      .then((result: any) => {
-        Toast('success', 'Thêm người vào phòng thành công');
-        router.push(`/manager/landlord/${id}/list-room`);
-      })
-      .catch((err) => {
-        console.log(err);
 
-        Toast('error', err.response.data.error);
-      });
-  };
   const onHandleRemove = async (id: any) => {
     // console.log(id);
+    const confirm = window.confirm("Bạn có chắc chắn muốn xóa không?")
+    if (confirm) {
+      await deleteBooking(id, userData)
+        .then((result: any) => {
+          setListBookings(listBookings.filter((item: { _id: any; }) => item._id !== id))
 
-    await deleteBooking(id, userData)
-      .then((result: any) => {
-        Toast('success', 'Xóa thành công');
-      })
-      .catch((err) => {
-        Toast('error', err.response.data.error);
-      });
+          Toast('success', 'Xóa thành công');
+        })
+        .catch((err) => {
+          Toast('error', err.response.data.error);
+        });
+    }
+
+  };
+  useEffect(() => {
+    const getListRoom = async () => {
+      const { data } = await listRoom(id, userData);
+      setListRooms(data.data);
+    };
+    getListRoom();
+  }, []);
+
+
+  const onSubmit = async (data1: any) => {
+    try {
+      const { data } = await createBooking(data1, userData, id);
+      const daata = data.data
+
+      setListBookings([...listBookings, daata])
+      setOpen(false)
+      Toast('success', 'Đặt tiền cọc thành công');
+
+    } catch (error: any) {
+      Toast('error', error.response.data.error);
+    }
   };
 
   return (
     <div>
+      <header className="bg-white shadow">
+        <div className="max-w-full mx-auto py-6 px-4 sm:px-6 lg:px-8">
+          <div className="lg:flex lg:items-center lg:justify-between">
+            <div className="flex-1 min-w-0">
+              <h2 className="text-2xl font-bold leading-7 text-gray-900 sm:text-2xl sm:truncate uppercase">
+                Cọc phòng
+              </h2>
+            </div>
+            <div className='text-right'>
+              <button
+                onClick={onOpenModal}
+                className="text-white bg-gradient-to-r from-cyan-500 to-blue-500 hover:bg-gradient-to-bl focus:ring-4 focus:outline-none focus:ring-cyan-300 dark:focus:ring-cyan-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2 "
+              >
+                Đặt cọc
+              </button>
+            </div>
+          </div>
+        </div>
+      </header>
       <div>
-        <button
-          onClick={onOpenModal}
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-full "
-        >
-          Đặt cọc
-        </button>
+
         <div>
           <div className="flex flex-col border bg-white mt-3">
             <div className="overflow-x-auto sm:-mx-6 lg:-mx-8">
@@ -91,9 +122,7 @@ const Booking = (props: Props) => {
                         <th scope="col" className="text-sm font-medium text-gray-900 px-6 py-4 text-left">
                           Số điện thoại
                         </th>
-                        <th scope="col" className="text-sm font-medium text-gray-900 px-6 py-4 text-left">
-                          Phòng
-                        </th>
+
                         <th scope="col" className="text-sm font-medium text-gray-900 px-6 py-4 text-left">
                           Tiền cọc
                         </th>
@@ -102,13 +131,14 @@ const Booking = (props: Props) => {
                         </th>
                       </tr>
                     </thead>
-                    <tbody>
-                      {listBookings.length > 0 ? (
-                        listBookings?.map((item: any, index: number) => {
-                          return (
-                            <>
+
+                    {listBookings && listBookings.length > 0 ? (
+                      listBookings?.map((item: any, index: number) => {
+                        return (
+                          <>
+                            <tbody>
                               <tr className="border-b">
-                                <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">1</td>
+                                <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{index+1}</td>
                                 <td className="text-sm text-gray-900 font-light px-6 py-4 whitespace-nowrap">
                                   {item.fullName}
                                 </td>
@@ -118,9 +148,7 @@ const Booking = (props: Props) => {
                                 <td className="text-sm text-gray-900 font-light px-6 py-4 whitespace-nowrap">
                                   {item.phoneNumber}
                                 </td>
-                                <td className="text-sm text-gray-900 font-light px-6 py-4 whitespace-nowrap">
-                                  {item.fullName}
-                                </td>
+
                                 <td className="text-sm text-gray-900 font-light px-6 py-4 whitespace-nowrap">
                                   {item.bookMoney}
                                 </td>
@@ -129,54 +157,29 @@ const Booking = (props: Props) => {
                                 </td>
                                 <td className="flex">
                                   <div>
-                                    <form action="" onSubmit={handleSubmit(onCreateRoom)}>
-                                      <div className="hidden">
-                                        <div>
-                                          {' '}
-                                          <input
-                                            type="text"
-                                            value={item._id}
-                                            {...register('idBooking', { required: true })}
-                                          />
-                                        </div>
-                                        <div>
-                                          {' '}
-                                          <input
-                                            type="text"
-                                            value={item.idRoom}
-                                            {...register('idRoom', { required: true })}
-                                          />
-                                        </div>
-                                      </div>
-                                      <div>
-                                        <button
-                                          type="submit"
-                                          className="focus:outline-none text-white bg-yellow-400 hover:bg-yellow-500 focus:ring-4 focus:ring-yellow-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:focus:ring-yellow-900"
-                                        >
-                                          Nhận phòng
-                                        </button>
-                                      </div>
-                                    </form>
+                                    {/* --------------------- */}
+                                    <AddBooking item1={item._id} item2={item.idRoom}></AddBooking>
                                   </div>
                                   <div>
-                                      
-                                      <button
-                                        type="submit"
-                                        className="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900"
-                                        onClick={() => onHandleRemove(item._id)}
-                                      >
-                                        Xóa
-                                      </button>
+
+                                    <button
+                                      type="submit"
+                                      className="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900"
+                                      onClick={() => onHandleRemove(item._id)}
+                                    >
+                                      Xóa
+                                    </button>
                                   </div>
                                 </td>
                               </tr>
-                            </>
-                          );
-                        })
-                      ) : (
-                        <div>Không có dữ liệu</div>
-                      )}
-                    </tbody>
+
+                            </tbody></>
+                        );
+                      })
+                    ) : (
+                      <div className='text-red-500 p-5'>Không có dữ liệu</div>
+                    )}
+
                   </table>
                 </div>
               </div>
@@ -195,7 +198,174 @@ const Booking = (props: Props) => {
                 <h2 className="pt-2 text-xl">Thông tin </h2>
               </div>
             </div>{' '}
-            <AddBooking close={onCloseModal}></AddBooking>
+            <div className="border mt-5 p-2">
+              <form className="w-full " onSubmit={handleSubmit(onSubmit)}>
+                <div className="md:flex md:items-center mb-6">
+                  <div className="md:w-1/5">
+                    <label
+                      className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                      htmlFor="inline-full-name"
+                    >
+                      Họ và tên
+                    </label>
+                  </div>
+                  <div className="md:w-2/3">
+                    <input
+                      className="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"
+                      id="inline-full-name"
+                      type="text"
+                      {...register('fullName', { required: true, minLength: 3 })}
+
+                    />
+                    <p className='text-red-500'>{errors.fullName?.type === "required" && <span>Không được đểtrống </span>}</p>
+                    <p className='text-red-500'>{errors.fullName?.type === "minLength" && <span>Tối thiểu 3 ký tự </span>}</p>
+                  </div>
+                </div>
+
+                <div className="md:flex md:items-center mb-6">
+                  <div className="md:w-1/5">
+                    <label
+                      className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                      htmlFor="inline-full-name"
+                    >
+                      Phòng
+                    </label>
+                  </div>
+                  <div className="relative">
+                    <select
+                      className="block appearance-none w-full bg-gray-200 border border-gray-200 text-gray-700 py-3 px-4 pr-8 rounded leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                      id="grid-state"
+                      {...register('idRoom')}
+                    >
+                      {listRooms &&
+                        listRooms.map((item: any, index: number) => {
+                          return (
+                            <>
+                              <option value={item._id}>{item.name}</option>
+                            </>
+                          );
+                        })}
+                    </select>
+                    <div className="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700">
+                      <svg className="fill-current h-4 w-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
+                        <path d="M9.293 12.95l.707.707L15.657 8l-1.414-1.414L10 10.828 5.757 6.586 4.343 8z" />
+                      </svg>
+                    </div>
+                  </div>
+                </div>
+                <div className="md:flex md:items-center mb-6 ">
+                  <div className="hidden">
+                    <div className="md:w-1/5">
+                      <label
+                        className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                        htmlFor="inline-password"
+                      >
+                        Nhà
+                      </label>
+                    </div>
+                    <div className="md:w-2/3">
+                      <input
+                        className="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"
+                        id="inline-password"
+                        type="text"
+                        value={id}
+                        {...register('idHouse', { required: true })}
+                      />
+                    </div>
+                  </div>
+                </div>
+
+                <div className="md:flex md:items-center mb-6">
+                  <div className="md:w-1/5">
+                    <label
+                      className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                      htmlFor="inline-password"
+                    >
+                      Email
+                    </label>
+                  </div>
+                  <div className="md:w-2/3">
+                    <input
+                      className="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"
+                      id="inline-password"
+                      type="email"
+                      {...register('email', { required: true })}
+                    />
+                    <p className='text-red-500'>{errors.email?.type === "required" && <span>Không được đểtrống </span>}</p>
+                  </div>
+                </div>
+                <div className="md:flex md:items-center mb-6">
+                  <div className="md:w-1/5">
+                    <label
+                      className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                      htmlFor="inline-full-name"
+                    >
+                      Số điện thoại
+                    </label>
+                  </div>
+                  <div className="md:w-2/3">
+                    <input
+                      className="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"
+                      id="inline-full-name"
+                      type="number"
+                      {...register('phoneNumber', { required: true, minLength: 10 })}
+                    />
+                    <p className='text-red-500'>{errors.phoneNumber?.type === "required" && <span>Không được đểtrống </span>}</p>
+                    <p className='text-red-500'>{errors.phoneNumber?.type === "minLength" && <span>Tối thiểu 10 ký tự </span>}</p>
+                  </div>
+                </div>
+                <div className="md:flex md:items-center mb-6">
+                  <div className="md:w-1/5">
+                    <label
+                      className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                      htmlFor="inline-full-name"
+                    >
+                      Tiền cọc
+                    </label>
+                  </div>
+                  <div className="md:w-2/3">
+                    <input
+                      className="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"
+                      id="inline-full-name"
+                      type="text"
+                      {...register('bookMoney', { required: true })}
+                    />
+                    <p className='text-red-500'>{errors.bookMoney?.type === "required" && <span>Không được đểtrống </span>}</p>
+                  </div>
+                </div>
+                <div className="md:flex md:items-center mb-6">
+                  <div className="md:w-1/5">
+                    <label
+                      className="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4"
+                      htmlFor="inline-full-name"
+                    >
+                      Thời gian
+                    </label>
+                  </div>
+                  <div className="md:w-2/3">
+                    <input
+                      className="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"
+                      id="inline-full-name"
+                      type="date"
+                      {...register('expectTime', { required: true })}
+                    />
+                    <p className='text-red-500'>{errors.expectTime?.type === "required" && <span>Không được đểtrống </span>}</p>
+
+                  </div>
+                </div>
+
+
+                <div className=" text-center">
+                  <button
+                    className="shadow bg-purple-500 hover:bg-purple-400 focus:shadow-outline focus:outline-none text-white font-bold py-2 px-4 rounded"
+                    type="submit"
+                  >
+                    Đặt cọc
+                  </button>
+                </div>
+
+              </form>
+            </div>
           </div>
         </Modal>
       </div>
